+++ packages/host/src/node/cli/bin.test.ts
@@ -2,6 +2,7 @@ import assert from "node:assert/strict";
 import { describe, it } from "node:test";
 import cp from "node:child_process";
 import path from "node:path";
+import { setupTempDirectory } from "../test-utils";
 
 const PACKAGE_ROOT = path.join(__dirname, "../../..");
 const BIN_PATH = path.join(PACKAGE_ROOT, "bin/react-native-node-api.mjs");
@@ -32,13 +33,28 @@ describe("bin", () => {
   });
 
   describe("link command", () => {
-    it("should succeed with a mention of Node-API modules", () => {
+    it("should succeed with a mention of Node-API modules", (context) => {
+      const targetBuildDir = setupTempDirectory(context, {});
+
       const { status, stdout, stderr } = cp.spawnSync(
         process.execPath,
-        [BIN_PATH, "link", "--android", "--apple"],
+        [
+          BIN_PATH,
+          "link",
+          "--android",
+          // Linking for Apple fails on non-Apple platforms
+          ...(process.platform === "darwin" ? ["--apple"] : []),
+        ],
         {
           cwd: PACKAGE_ROOT,
           encoding: "utf8",
+          env: {
+            ...process.env,
+            TARGET_BUILD_DIR: targetBuildDir,
+            FRAMEWORKS_FOLDER_PATH: "Frameworks",
+            PLATFORM_NAME: "iphonesimulator",
+            ARCHS: "arm64",
+          },
         },
       );
